Close #06 : Authorization
- Change role access in security config
- Only users with the ADMIN role can access the user list
- Adding a personal access denied message
- Only users who have created a task can delete it
- Only users with the ADMIN role can deleted a task with anonymous author
- Add test for testing different permissions

+++ src/AppBundle/Controller/TaskController.php
@@ -88,11 +88,18 @@ class TaskController extends Controller
      */
     public function deleteTaskAction(Task $task)
     {
-        $em = $this->getDoctrine()->getManager();
-        $em->remove($task);
-        $em->flush();
+        if ($this->getUser() === $task->getAuthor() || $this->isGranted('ROLE_ADMIN') && !$task->asAuthor()) {
 
-        $this->addFlash('success', 'La tâche a bien été supprimée.');
+            $em = $this->getDoctrine()->getManager();
+            $em->remove($task);
+            $em->flush();
+
+            $this->addFlash('success', 'La tâche a bien été supprimée.');
+
+        } else {
+
+            $this->addFlash('error', 'Vous ne pouvez pas supprimer un tâche d\'un autre auteur.');
+        }
 
         return $this->redirectToRoute('task_list');
     }
